# openapi.yaml
openapi: 3.1.0
info:
  title: Devlogia Public API
  version: 1.0.0
  description: OpenAPI definition covering the public content and analytics endpoints exposed by Devlogia. Authentication
    is handled via NextAuth session cookies.
  contact:
    name: Devlogia Engineering
    url: https://devlogia.test
    email: engineering@devlogia.test
servers:
  - url: https://devlogia.test
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - CookieAuth: []
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: Session token issued by NextAuth. In browsers it is managed automatically after sign-in.
  schemas:
...
    UploadFile:
      type: object
      properties:
        id:
          type: string
          example: media_123
        url:
          type: string
          format: uri
          example: https://cdn.devlogia.test/uploads/2024-04-01/header.png
        path:
          type: string
          example: uploads/2024-04-01/header.png
        mimeType:
          type: string
          example: image/png
        sizeBytes:
          type: number
          example: 24576
        checksum:
          type: string
          minLength: 64
          maxLength: 64
          example: 0f343b0931126a20f133d67c2b018a3b
        alt:
          type: string
          example: Team photo
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - url
        - path
        - mimeType
        - sizeBytes
        - checksum
        - alt
        - createdAt
    UploadResponse:
      type: object
      properties:
        success:
          type: boolean
        provider:
          type: string
          enum:
            - supabase
            - stub
        files:
          type: array
          items:
            $ref: "#/components/schemas/UploadFile"
      required:
        - success
        - provider
        - files
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Unauthorized
      required:
        - error
...
paths:
...
  /api/uploadthing:
    post:
      tags:
        - Uploads
      summary: Upload media to the configured storage provider
      security:
        - CookieAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File contents
              required:
                - file
      responses:
        "200":
          description: Upload succeeded (returns stub provider when S3 isn't configured).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...